ref: convert brightness values at a lower level and scan only when vm initializes

+++ app/src/main/java/com/example/lumen/data/mapper/LedControllerMapper.kt
@@ -13,7 +13,7 @@ fun ByteArray.toLedControllerState(): LedControllerState {
     val isOn = this[0] == 0x01.toByte()
     val preset = this[1]
     val speed = this[2]
-    val brightness = this[3].toUByte().toInt()
+    val brightness = this[3].toUByte().toFloat()
     val icMode = this[4]
     val channel = this[5]
     val pixelCountMSB = this[6].toUByte().toInt()
